Update README.md
Since I often look for specific snippets in my scripts and no longer know where I used them,
I wrote a small tool that searches for specific keywords in any file.

The program searches for a specific word in a folder in all files and subfolders.
The result is written in a table. The result can then be opened with a mouse click.
The search can also be narrowed down with filters.